Edits to README
- Edited the {dm} README file for language and clairity.

- Created a new file README_brief as a possible alternative.

That file excludes the Good-to-know section of README about relational DB concepts because the same information is included verbatim in the intro-to-relational-data-models vignette.

The RDMS section is replaced with a more general Background section that explains the problem that {dm} attempts to solve. Maybe it  can be shortedned as well.  It is best if a README file presents only the bare minimum for users to get started with a package.

Please comment!